refactor: move UserService.java dedicated feature service
